refactor fetchers, add utils file for shared methods, add tests
